Command first pass honor the `--api-version` passed in
If the `--api-version` exists for the command, read that versioned command's metadata to construct the command. Otherwise, return the latest versioned command's metadata to build the command, which will then fail with the reason that the specified api version doesn't exist.

The benefit of doing this is to ensure the help message and the command metadata used for the invocation is using the right version.

Fix #21